+++ R/chunk_images.R
@@ -1,3 +1,18 @@
+# note_docx_images -----
+
+#' @title internal utils for roxygen tags reuse
+#' @note
+#' This chunk option requires package officedown in a R Markdown
+#' context with Word output format. With Quarto (`format: docx`) or
+#' `rmarkdown::word_document()`, the resulting file must be repaired
+#' with [repair_docx()].
+#'
+#' PowerPoint cannot mix images and text in a paragraph, images
+#' are removed when outputing to PowerPoint format.
+#' @name note_docx_images
+#' @keywords internal
+NULL
+
 #' @importFrom grDevices as.raster
 #' @export
 #' @title Image chunk
@@ -22,12 +37,7 @@
 #' @param alt alternative text for the image (used for accessibility)
 #' @param ... unused argument
 #' @family chunk elements for paragraph
-#' @note
-#' This chunk option requires package officedown in a R Markdown
-#' context with Word output format.
-#'
-#' PowerPoint cannot mix images and text in a paragraph, images
-#' are removed when outputing to PowerPoint format.
+#' @inherit note_docx_images note
 #' @seealso [compose()], [as_paragraph()]
 #' @examples
 #' img.file <- file.path(
@@ -109,12 +119,7 @@ as_image <- function(
 #' @param width,height size of the resulting png file in inches
 #' @param unit unit for width and height, one of "in", "cm", "mm".
 #' @param alt alternative text for the image (used for accessibility)
-#' @note
-#' This chunk option requires package officedown in a R Markdown
-#' context with Word output format.
-#'
-#' PowerPoint cannot mix images and text in a paragraph, images
-#' are removed when outputing to PowerPoint format.
+#' @inherit note_docx_images note
 #' @family chunk elements for paragraph
 #' @examples
 #' ft <- flextable(head(iris, n = 10))
@@ -211,12 +216,7 @@ minibar <- function(
 #' @param alt alternative text for the image (used for accessibility)
 #' @param raster_width number of pixels used as width
 #' when interpolating value.
-#' @note
-#' This chunk option requires package officedown in a R Markdown
-#' context with Word output format.
-#'
-#' PowerPoint cannot mix images and text in a paragraph, images
-#' are removed when outputing to PowerPoint format.
+#' @inherit note_docx_images note
 #' @family chunk elements for paragraph
 #' @examples
 #' myft <- flextable(head(iris, n = 10))
@@ -325,12 +325,7 @@ linerange <- function(
 #' @param unit unit for width and height, one of "in", "cm", "mm".
 #' @param alt alternative text for the image (used for accessibility)
 #' @param ... arguments sent to plot functions (see [par()])
-#' @note
-#' This chunk option requires package officedown in a R Markdown
-#' context with Word output format.
-#'
-#' PowerPoint cannot mix images and text in a paragraph, images
-#' are removed when outputing to PowerPoint format.
+#' @inherit note_docx_images note
 #' @family chunk elements for paragraph
 #' @examples
 #' library(data.table)
@@ -471,12 +466,7 @@ plot_chunk <- function(
 #' @param unit unit for width and height, one of "in", "cm", "mm".
 #' @param res resolution of the png image in ppi
 #' @param alt alternative text for the image (used for accessibility)
-#' @note
-#' This chunk option requires package officedown in a R Markdown
-#' context with Word output format.
-#'
-#' PowerPoint cannot mix images and text in a paragraph, images
-#' are removed when outputing to PowerPoint format.
+#' @inherit note_docx_images note
 #' @family chunk elements for paragraph
 #' @examples
 #' library(data.table)
@@ -570,12 +560,7 @@ gg_chunk <- function(
 #' @param unit unit for width and height, one of "in", "cm", "mm".
 #' @param res resolution of the png image in ppi
 #' @param alt alternative text for the image (used for accessibility)
-#' @note
-#' This chunk option requires package officedown in a R Markdown
-#' context with Word output format.
-#'
-#' PowerPoint cannot mix images and text in a paragraph, images
-#' are removed when outputing to PowerPoint format.
+#' @inherit note_docx_images note
 #' @family chunk elements for paragraph
 #' @examples
 #' library(flextable)